services/accounts/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
gunicorn==22.0.0
dj-database-url==2.2.0
django-cors-headers==3.10.0
psycopg2-binary==2.9.9
whitenoise==6.6.0